Make check_procs cgroups aware.
This patch enables check_proc to monitor processes in PID name-
spaced environments.

Normally, all processes are visible from host - the ones running
directly on host and the ones in containers. This results in
check_procs reporting false-positives or false negatives depending
on whether he sees to many of given processes on to little.

These changes introduce one filter to check_proc - cgroup hierarchy.
Specified on the command line, it allows user to define which cgroup
hierarchy we are interested in and check_proc is ignoring all other.

Change-Id: Icbed79b6bad45dc387e087cf48baf2715ccda430