plugins/check_http.c - fixes and cleanup, chunking
This corrects header issues from the use of sizeof(), and thus causing the string "Transfer-Encoding: chunked" to enter the while loops at 745 and 746 looking like "-Encoding: chunked", which did not match the check for only inspace or :.

Also minor corrections for our coding standards. This will all get reformatted shortly, so we might as well start now.