check_dns.c - macro querytypes and auto cnames
Most of the code for checking querytypes can be abstracted into a single macro with a few variables. This significantly reduces code and complexity of potential issues. This also corrects a non-null terminated string possibility. Finally we are checking the set querytype and allowing drop through to the accept_cname statement when the user is querying for cnames as the logic depends on BOTH being set. This avoids them having to remember to set both, as that is a bit absurd on my part.