check_dns.c - multiple fixes
coverity 103711 - increase query_type size to avoid non-null termination.

coverity 103719 - avoid usage of potentially null pointer, which may not be null as we ++ it prior to checking or use.

covierity 103721 - Corrected issue with previous commit where nslookup results were not being processed beyond the first line due to me incorrectly believing result was set to OK when initially checked. This still preserves the same logic to keep states as they worsen throughout the plugins logic.

Added tmp_dns_server for when -s x.x.x.x is not provided and we die within error scan. This will be much more informative and provide the server that is resulting in an error state. This does not conflict with instances where -s was provided and we should only be checking one server or logic where that server should be provided in output.

Removed leading spaces from when addresses were strduped into addresses array, causing incorrect output.