Use strlen instead of sizeof in check_dns
The value given to check_dns' "-q"/"--querytype" option is converted to
uppercase. The optarg variable is of type char*, requiring the use of
strlen(). sizeof() of a char* returns the platform's pointer size
(usually 4 or 8 bytes). If the value given to "-q" is shorter, the code
ends up overwriting unrelated memory:

  $ check_dns -H nagios.com -q a --expected-address 1.2.3.4
  check_dns: unrecognized option '--EXPEcted-address'