check_http: -e breaks -f
Fix for issue https://github.com/nagios-plugins/nagios-plugins/issues/91

Checking status codes and following redirects makes for compicated
coding and testing. But this change seems to handle all situations
properly.

It is important to note, though, that anything specified by `-e`
will only be compared agains the final page. So if you use `-e 301`
expecting a redirect, it will error out because the status of the
last page would be `200` (or some other status).