check_dns: Segfaulting with timeout > 26 sec
Fix for issue #161

`nslookup` has it's own timeout, then it exits with the message
`;; connection timed out; no servers could be reached`.
`check_dns` then tries to parse `stdout` looking for a `:`
character, which it doesn't find, so `msg` gets set to NULL.
Then it tries to print that message and segfaults.