check_disk: alerts issued too soon
Fix for issue #195

The percentage calculated for used space does effectively a ceil(p)
call. So if used space is 89.01% it would get set to 90%. Free space
percentage is `100 - usedspace%`. With a warning threshold of 10%,
a warning would be triggered when there was 10.9% free. So on a
partition of 100GB, a warning would be triggered 100MB before the
actual threshold.

I changed the used space percentage calculation to calculate to
two decimal points, which prevents a warning or critical condition
from being triggered until much closer to the specified thresholds.