check_file_age: fix check for multiple files
Without this, the script would always return the result for the last
checked file, if we check multiple files (e.g a directory).

We now return WARNING/CRITICAL for the first file which violates the
thresholds, and returnung the perf data for this file.

If all files are OK, we return only OK with the perf data of the first
file found. I don't added the filenames here, because if we check a
directory with many files, it don't makes sense to me to put all filename
in the check results.