handle errors in "grep -c"
This script returns with critical status and message "()" when the previous "diff" or "grep" fails. 
Because the case, if $count is not numeric (e.g. empty), will not be handled correctly.

This fix handle errors in "grep -c", e.g. if the file not exists, and return with Unknown status and stderr message.

Example
Log check error: grep: can't open /tmp/check_log.002330