Resolve #507 - stop reading when Content-Length matches the amount of content received
Prior to this commit, check_http relied on read() returning 0 or erroring in order to
exit the loop. This commit keeps the same behavior, but once headers have been received,
it also reads the Content-Length and compares it to the number of bytes received after
the start of the body.