Add option to output response body as performance data
Added option -o/--output-body-as-perfdata to append the http response body
to the output line. This will only happen for responses that are of state "OK".

Imagine the requested webserver responds with 200 OK and the text:

|'first variable'=42|'second variable'=111

This text will be appended to the plugin output, and is treated as regular
performance data.

The switch -o or --output-body-as-perfdata enables this feature.