Replace 0 with U in performance output.
When packet loss is 100%, then rta, rtmax, rtmin, jitter, jitter_max and jitter_min are set to 0ms in performance output. This of course is wrong: these values are unknown in this case. The Nagios plugin development guide at https://nagios-plugins.org/doc/guidelines.html#AEN200 allows to use "U" for unknown values.
I think we should output U instead of 0 in performance output when there is no reply from the target. U (Unknown) is a lot better than 0ms, which is pretty quick for a host that is down.